% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/another_function.R
\name{another_function}
\alias{another_function}
\title{another_function}
\usage{
another_function(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
another_function
}
\examples{
x <- c(1,2,3,4)
another_function(x)
}
