% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{new_function}
\alias{new_function}
\title{new_function}
\usage{
new_function(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a numeric as well
}
\description{
new_function
}
\examples{
new_function(6)
}
