% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third_function.R
\name{third_function}
\alias{third_function}
\title{third_function}
\usage{
third_function(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
third_function
}
\examples{
x <- c(1,2,3,4)
third_function(x)
}
